update batch_valid_length block_tables dtype int32

+++ mindformers/generation/text_generator.py
@@ -310,8 +310,8 @@ class GenerationMixin:
             self.add_flags_recursive(is_first_iteration=True)
             model_inputs["input_position"] = Tensor(current_index, mstype.int32)
             model_inputs["init_reset"] = Tensor([False], mstype.bool_)  # init_reset (1,) bool False
-            model_inputs["batch_valid_length"] = Tensor([valid_length_each_example], mstype.int64)
-            model_inputs["block_tables"] = Tensor(block_tables, mstype.int64)
+            model_inputs["batch_valid_length"] = Tensor([valid_length_each_example], mstype.int32)
+            model_inputs["block_tables"] = Tensor(block_tables, mstype.int32)
             model_inputs["slot_mapping"] = Tensor(slot_mapping, mstype.int32)
             # pylint: disable=E1102
             res = self(
@@ -327,8 +327,8 @@ class GenerationMixin:
             self.slice_incremental_inputs(model_inputs, current_index)
             model_inputs["input_position"] = Tensor(current_index, mstype.int32)
             model_inputs["init_reset"] = Tensor([True], mstype.bool_)  # init_reset (1,) bool True
-            model_inputs["batch_valid_length"] = Tensor([valid_length_each_example], mstype.int64)
-            model_inputs["block_tables"] = Tensor(block_tables, mstype.int64)
+            model_inputs["batch_valid_length"] = Tensor([valid_length_each_example], mstype.int32)
+            model_inputs["block_tables"] = Tensor(block_tables, mstype.int32)
             model_inputs["slot_mapping"] = Tensor(slot_mapping, mstype.int32)
             # pylint: disable=E1102
             res = self(
